code refactor send funds
